requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.6
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.2.1
